requirements.txt
docling
safetensors
certifi
requests
token_count
openai
python-dotenv
